requirements.txt
Flask
requests
beautifulsoup4
DateTime
line-bot-sdk
Gunicorn